requirements.txt
# Core
numpy>=1.26.0
pandas>=2.0.0

# Computer Vision
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ML / AI (demo + future real model)
torch>=2.0.0
torchvision>=0.17.0
ultralytics>=8.2.0

# App / UI
streamlit>=1.28.0

# Utils
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.66.0

# Testing (agentic spike)
pytest>=7.4.0
httpx>=0.27.0
python-dotenv>=1.0.0
